Realizado alteração em algumas tabelas e criado o seed de um album

+++ database/migrations/2020_06_19_230948_create_album_page_photos_table.php
@@ -18,9 +18,9 @@ class CreateAlbumPagePhotosTable extends Migration
             $table->unsignedBigInteger('page_id');
             $table->unsignedBigInteger('frame_id');
             $table->integer('sequence');
-            $table->integer('x_position');
-            $table->integer('y_position');
-            $table->integer('rotation');
+            $table->decimal('x_position');
+            $table->decimal('y_position');
+            $table->decimal('rotation');
             $table->timestamps();
         });
     }
